Refactored device slice into new UI slice
