+++ src/components/Map/AnalyticsPane.tsx
@@ -12,14 +12,14 @@ import {
   requestRunAllAlgorithms,
 } from "@features/algorithms/algorithmsActions";
 import { selectAlgorithmsResults } from "@features/algorithms/algorithmsSelectors";
-import { deviceSliceActions } from "@features/device/deviceSlice";
+import { uiSliceActions } from "@features/ui/slice";
 
 const AnalyticsPane = () => {
   const dispatch = useDispatch();
   const { apResult, mincutResult } = useSelector(selectAlgorithmsResults());
 
   const handleClosePane = () => {
-    dispatch(deviceSliceActions.setInfoPane(null));
+    dispatch(uiSliceActions.setInfoPane(null));
   };
 
   const diffcen = new Map();
