+++ src/components/Map/MapNodeTooltip.tsx
@@ -1,10 +1,9 @@
-import {
-  selectNodeById,
-  selectRootState,
-} from "@app/features/device/deviceSelectors";
 import type { PickingInfo } from "deck.gl/typed";
 import React from "react";
 import { useSelector } from "react-redux";
+
+import { selectNodeById } from "@features/device/deviceSelectors";
+import { selectRootState } from "@features/ui/selectors";
 
 export interface IMapNodeTooltip {
   hoverInfo: PickingInfo;
