+++ src/components/NodeSearch/NodeSearchDock.tsx
@@ -13,14 +13,11 @@ import DefaultTooltip from "@components/DefaultTooltip";
 import NodeSearchInput from "@components/NodeSearch/NodeSearchInput";
 import NodeSearchResult from "@components/NodeSearch/NodeSearchResult";
 
-import {
-  selectActiveNodeId,
-  selectDevice,
-  selectAllNodes,
-} from "@features/device/deviceSelectors";
-import { deviceSliceActions } from "@features/device/deviceSlice";
+import { selectDevice, selectAllNodes } from "@features/device/deviceSelectors";
 import { selectMapUIState } from "@features/map/mapSelectors";
 import { mapSliceActions } from "@features/map/mapSlice";
+import { selectActiveNodeId } from "@features/ui/selectors";
+import { uiSliceActions } from "@features/ui/slice";
 
 import { MapIDs, getFlyToConfig } from "@utils/map";
 
@@ -101,7 +98,7 @@ const NodeSearchDock = () => {
 
   const handleNodeSelect = (nodeId: number) => {
     const isNodeActive = activeNodeId === nodeId;
-    dispatch(deviceSliceActions.setActiveNode(isNodeActive ? null : nodeId));
+    dispatch(uiSliceActions.setActiveNode(isNodeActive ? null : nodeId));
 
     // Only animate when node is not currently active
     if (isNodeActive) return;
