+++ src/features/connection/connectionSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import type { RequestStatus } from "@features/requests/requestReducer";
+import type { RequestStatus } from "@features/requests/requestSlice";
 import type { DeviceKey } from "@utils/connections";
 
 export interface IConnectionState {
