+++ tsconfig.json
@@ -1,9 +1,16 @@
 {
-  "include": ["src", "types"],
+  "include": [
+    "src",
+    "types"
+  ],
   "compilerOptions": {
     "target": "ESNext",
     "useDefineForClassFields": true,
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
+    "lib": [
+      "DOM",
+      "DOM.Iterable",
+      "ESNext"
+    ],
     "allowJs": false,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -18,18 +25,33 @@
     "jsx": "react-jsx",
     "baseUrl": "./",
     "paths": {
-      "@app/*": ["src/*"],
-      "@bindings/*": ["src/bindings/*"],
-      "@components/*": ["src/components/*"],
-      "@features/*": ["src/features/*"],
-      "@store/*": ["src/store/*"],
-      "@utils/*": ["src/utils/*"]
+      "@bindings/*": [
+        "src/bindings/*"
+      ],
+      "@components/*": [
+        "src/components/*"
+      ],
+      "@features/*": [
+        "src/features/*"
+      ],
+      "@store/*": [
+        "src/store/*"
+      ],
+      "@utils/*": [
+        "src/utils/*"
+      ],
+      "@app/*": [
+        "src/*"
+      ],
     },
     "importHelpers": true,
     "removeComments": true,
     "strictNullChecks": true,
-    "types": ["vite/client", "node", "jest"],
-    "importsNotUsedAsValues": "error",
+    "types": [
+      "vite/client",
+      "node",
+      "jest"
+    ],
     "strictPropertyInitialization": false,
     "experimentalDecorators": true
   }
